Добавил триггеры для действий плагинов при удалении пользователя слогин и пользователя джумлы. Доработал плагин аватаров для удаления аватаровпри удалени пользователей.

+++ com_slogin/site/controller.php
@@ -206,13 +206,30 @@ class SLoginController extends SLoginControllerParent
 
     private function deleteSloginUser($slogin_id, $provider){
         $db = JFactory::getDbo();
+
+        JPluginHelper::importPlugin('slogin_integration');
+        $dispatcher = JDispatcher::getInstance();
+        $query = $db->getQuery(true);
+        $query->select('id');
+        $query->from('#__slogin_users');
+        $query->where($db->quoteName('slogin_id') . ' = ' . $db->quote($slogin_id));
+        $query->where($db->quoteName('provider') . ' = ' . $db->quote($provider));
+        $db->setQuery((string)$query, 0, 1);
+        $id = (int)$db->loadResult();
+
+        if($id == 0){
+            return false;
+        }
+
+        $dispatcher->trigger('onBeforeSloginDeleteSloginUser',array($id));
         $query = $db->getQuery(true);
         $query->delete();
         $query->from($db->quoteName('#__slogin_users'));
-        $query->where($db->quoteName('slogin_id') . ' = ' . $db->quote($slogin_id));
-        $query->where($db->quoteName('provider') . ' = ' . $db->quote($provider));
+        $query->where($db->quoteName('id') . ' = ' . $db->quote($id));
         $db->setQuery($query);
         $db->query();
+        $dispatcher->trigger('onAfterSloginDeleteSloginUser',array($id));
+        return true;
     }
 
     /**
@@ -695,7 +712,7 @@ class SLoginController extends SLoginControllerParent
                 }
                 else
                 {
-                $return = base64_decode($model->getReturnURL($this->config, 'after_reg_redirect'));
+                    $return = base64_decode($model->getReturnURL($this->config, 'after_reg_redirect'));
                 }
 
 
@@ -753,12 +770,14 @@ class SLoginController extends SLoginControllerParent
 
         $db = JFactory::getDbo();
         $query = $db->getQuery(true);
-        $query->delete();
+        $query->select($db->quoteName('slogin_id'));
         $query->from($db->quoteName('#__slogin_users'));
         $query->where($db->quoteName('user_id') . ' = ' . $db->quote($user_id));
         $query->where($db->quoteName('provider') . ' = ' . $db->quote($provider));
-        $db->setQuery($query);
-        $db->query();
+        $db->setQuery($query, 0, 1);
+        $slogin_id = $db->loadResult();
+
+        $this->deleteSloginUser($slogin_id, $provider);
 
         $this->displayRedirect($link);
     }
